+++ io-service/src/main/java/com/example/ioservice/model/UserModel.java
@@ -1,0 +1,23 @@
+package com.example.ioservice.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name= "users")
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class UserModel {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long userId;
+    private String username;
+    private String email;
+    private String address;
+    private String role;
+}