+++ io-service/src/main/java/com/example/ioservice/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.example.ioservice.repository;
+
+import com.example.ioservice.model.UserModel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserModel, Long> {
+    UserModel findByUsername(String username);
+}