+++ market-service/src/main/java/com/example/marketservice/dto/ProductDto.java
@@ -1,0 +1,4 @@
+package com.example.marketservice.dto;
+
+public record ProductDto(Long productId, String productName, Float price, String description, Long ownerId, Integer quantity) {
+}